1차 마무리

+++ app/src/main/java/com/example/aop_part4_chapter03/model/LocationLatLngEntity.kt
@@ -5,6 +5,6 @@ import kotlinx.android.parcel.Parcelize
 
 @Parcelize
 data class LocationLatLngEntity(
-    var latitude : Float,
-    var longtitude : Float
+    var latitude : Double,
+    var longtitude : Double
 ) : Parcelable
